3.0.0.610

+++ src/oscontrol/deepinabilitymanager.h
@@ -113,7 +113,7 @@ public:
     OSCallContext doSeek(int offset);
 
     // 系统字号
-    OSCallContext doSystemFontSize(float size);
+    OSCallContext doSystemFontSize(int size);
     OSCallContext getSystemFontSize();
 
     // 文件操作
